requirements.txt
edge-tts>=7.0.0
customtkinter>=5.2.0
piper-tts>=1.2.0
